modularize logical ops
