+++ tfjs-core/src/gradients/Prelu_grad.ts
@@ -19,10 +19,10 @@ import {GradConfig} from '../kernel_registry';
 import {reshape} from '../ops/array_ops';
 import {getReductionAxes} from '../ops/broadcast_util';
 import {greater} from '../ops/greater';
-import {where} from '../ops/logical_ops';
 import {mul} from '../ops/mul';
 import {sum} from '../ops/reduction_ops';
 import {zerosLike} from '../ops/tensor_ops';
+import {where} from '../ops/where';
 import {Tensor} from '../tensor';
 
 export const preluGradConfig: GradConfig = {
