+++ tfjs-core/src/ops/ops.ts
@@ -58,6 +58,10 @@ export {leakyRelu} from './leaky_relu';
 export {less} from './less';
 export {lessEqual} from './less_equal';
 export {localResponseNormalization} from './local_response_normalization';
+export {logicalAnd} from './logical_and';
+export {logicalNot} from './logical_not';
+export {logicalOr} from './logical_or';
+export {logicalXor} from './logical_xor';
 export {matMul} from './mat_mul';
 export {max} from './max';
 export {maxPool} from './max_pool';
@@ -95,6 +99,8 @@ export {squaredDifference} from './squared_difference';
 export {sub} from './sub';
 export {tile} from './tile';
 export {truncatedNormal} from './truncated_normal';
+export {where} from './where';
+export {whereAsync} from './where_async';
 
 export * from './boolean_mask';
 export * from './reverse';
@@ -103,7 +109,6 @@ export * from './unary_ops';
 export * from './reduction_ops';
 export * from './compare';
 export * from './binary_ops';
-export * from './logical_ops';
 export * from './array_ops';
 export * from './tensor_ops';
 export * from './transpose';
