+++ tfjs-core/src/ops/where_async.ts
@@ -1,0 +1,52 @@
+/**
+ * @license
+ * Copyright 2020 Google Inc. All Rights Reserved.
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * =============================================================================
+ */
+import {whereImpl} from '../backends/where_impl';
+import {Tensor, Tensor2D} from '../tensor';
+import {convertToTensor} from '../tensor_util_env';
+import {TensorLike} from '../types';
+
+/**
+ * Returns the coordinates of true elements of condition.
+ *
+ * The coordinates are returned in a 2-D tensor where the first dimension (rows)
+ * represents the number of true elements, and the second dimension (columns)
+ * represents the coordinates of the true elements. Keep in mind, the shape of
+ * the output tensor can vary depending on how many true values there are in
+ * input. Indices are output in row-major order. The resulting tensor has the
+ * shape `[numTrueElems, condition.rank]`.
+ *
+ * This is analogous to calling the python `tf.where(cond)` without an x or y.
+ *
+ * ```js
+ * const cond = tf.tensor1d([false, false, true], 'bool');
+ * const result = await tf.whereAsync(cond);
+ * result.print();
+ * ```
+ */
+/** @doc {heading: 'Operations', subheading: 'Logical'} */
+async function whereAsync_(condition: Tensor|TensorLike): Promise<Tensor2D> {
+  const $condition =
+      convertToTensor(condition, 'condition', 'whereAsync', 'bool');
+  const vals = await $condition.data();
+  const res = whereImpl($condition.shape, vals);
+  if (condition !== $condition) {
+    $condition.dispose();
+  }
+  return res;
+}
+
+export const whereAsync = whereAsync_;
