+++ tfjs-core/src/public/chained_ops/register_all_chained_ops.ts
@@ -40,6 +40,10 @@ import './leaky_relu';
 import './less';
 import './less_equal';
 import './local_response_normalization';
+import './logical_and';
+import './logical_not';
+import './logical_or';
+import './logical_xor';
 import './mat_mul';
 import './max';
 import './max_pool';
@@ -65,3 +69,4 @@ import './space_to_batch_nd';
 import './sub';
 import './tile';
 import './transpose';
+import './where';
