+++ tfjs-core/src/register_all_gradients.ts
@@ -50,6 +50,7 @@ import {relu6GradConfig} from './gradients/Relu6_grad';
 import {reluGradConfig} from './gradients/Relu_grad';
 import {resizeBilinearGradConfig} from './gradients/ResizeBilinear_grad';
 import {resizeNearestNeighborGradConfig} from './gradients/ResizeNearestNeighbor_grad';
+import {selectV2PoolGradConfig} from './gradients/SelectV2_grad';
 import {seluGradConfig} from './gradients/Selu_grad';
 import {spaceToBatchNDGradConfig} from './gradients/SpaceToBatchND_grad';
 import {splitVGradConfig} from './gradients/SplitV_grad';
@@ -105,6 +106,7 @@ const gradConfigs: GradConfig[] = [
   resizeNearestNeighborGradConfig,
   relu6GradConfig,
   seluGradConfig,
+  selectV2PoolGradConfig,
   spaceToBatchNDGradConfig,
   splitVGradConfig,
   squareGradConfig,
