+++ tfjs-core/src/tests.ts
@@ -84,7 +84,10 @@ import './ops/less_equal_test';
 import './ops/less_test';
 import './ops/linalg_ops_test';
 import './ops/local_response_normalization_test';
-import './ops/logical_ops_test';
+import './ops/logical_and_test';
+import './ops/logical_not_test';
+import './ops/logical_or_test';
+import './ops/logical_xor_test';
 import './ops/loss_ops_test';
 import './ops/lstm_test';
 import './ops/mat_mul_test';
@@ -125,6 +128,8 @@ import './ops/topk_test';
 import './ops/transpose_test';
 import './ops/truncated_normal_test';
 import './ops/unary_ops_test';
+import './ops/where_async_test';
+import './ops/where_test';
 import './optimizers/adadelta_optimizer_test';
 import './optimizers/adagrad_optimizer_test';
 import './optimizers/adam_optimizer_test';
